feat[step-2]: refactoring]
