+++ src/errors/http-error/types.ts
